fix : separate add annotator script #13
